Move the user agent property from the order to the conversion. This simplifies things
