+++ src/Factory/ConversionFactory.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Setono\SyliusGoogleAdsPlugin\Factory;
 
 use Setono\SyliusGoogleAdsPlugin\Model\ConversionInterface;
-use Setono\SyliusGoogleAdsPlugin\Model\OrderInterface;
+use Sylius\Component\Core\Model\OrderInterface;
 use Sylius\Component\Resource\Factory\FactoryInterface;
 use Webmozart\Assert\Assert;
 
@@ -34,11 +34,7 @@ final class ConversionFactory implements ConversionFactoryInterface
         $conversion->setValue($order->getTotal());
         $conversion->setCurrencyCode((string) $order->getCurrencyCode());
         $conversion->setOrder($order);
-
-        $clickId = $order->getGoogleClickId();
-        if (null !== $clickId) {
-            $conversion->setGoogleClickId($clickId);
-        }
+        $conversion->setChannel($order->getChannel());
 
         return $conversion;
     }
