+++ src/Resources/config/services.xml
@@ -12,7 +12,6 @@
         <import resource="services/form.xml"/>
         <import resource="services/menu.xml"/>
         <import resource="services/message.xml"/>
-        <import resource="services/order_processor.xml"/>
         <import resource="services/provider.xml"/>
         <import resource="services/resolver.xml"/>
     </imports>
