--- src/Resources/config/services/order_processor.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-
-<container xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://symfony.com/schema/dic/services"
-           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
-    <services>
-        <service id="setono_sylius_google_ads.order_processor.set_gclid_and_user_agent"
-                 class="Setono\SyliusGoogleAdsPlugin\OrderProcessor\SetGclidAndUserAgentOrderProcessor">
-            <argument type="service" id="request_stack"/>
-            <argument>%setono_sylius_google_ads.gclid_cookie_name%</argument>
-
-            <tag name="sylius.order_processor"/>
-        </service>
-    </services>
-</container>
