+++ tests/Application/config/packages/_sylius.yaml
@@ -7,12 +7,6 @@ imports:
 parameters:
     sylius_core.public_dir: '%kernel.project_dir%/public'
 
-sylius_order:
-    resources:
-        order:
-            classes:
-                model: Tests\Setono\SyliusGoogleAdsPlugin\Application\Model\Order
-
 sylius_shop:
     product_grid:
         include_all_descendants: true
