+++ tests/Application/config/packages/doctrine.yaml
@@ -12,13 +12,3 @@ doctrine:
         charset: UTF8
 
         url: '%env(resolve:DATABASE_URL)%'
-    orm:
-        auto_generate_proxy_classes: '%kernel.debug%'
-        auto_mapping: true
-        mappings:
-            App:
-                is_bundle: false
-                type: annotation
-                dir: '%kernel.project_dir%/Model'
-                prefix: 'Tests\Setono\SyliusGoogleAdsPlugin\Application\Model'
-                alias: App
